separated out presenations logic & added map toggle
