+++ src/pages/dashboard/.Map.scss
@@ -1,0 +1,27 @@
+%map-toggle-active {
+  color: black;
+  border-color: black;
+  font-weight: 500;
+}
+%map-toggle-inactive {
+  color: lightgray;
+  border: solid lightgray 2px;
+  font-weight: initial;
+}
+
+#map-style-button span {
+  @extend %map-toggle-inactive;
+  display: inline-block;
+  margin: -1px;
+  padding: 2px 6px;
+  &:nth-child(1) {
+    border-radius: 8px 0 0 8px;
+  }
+  &:nth-child(2) {
+    border-radius: 0 8px 8px 0;
+    border-left-color: black !important;
+  }
+  &.active {
+    @extend %map-toggle-active;
+  }
+}
