+++ src/pages/dashboard/charts/map.scss
