+++ src/pages/dashboard/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Layout from '../../components/layout'
-import Map from './_Map'
+import Map from './.Map'
 
 const DashboardPage = () => (
   <Layout title="Visual Dashboard">
